app/requirements.txt
aniso8601==9.0.1
attrs==21.2.0
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
Flask==2.0.2
Flask-RESTful==0.3.9
flask-restx==0.5.1
idna==3.3
install==1.3.4
itsdangerous==2.0.1
Jinja2==3.0.2
jsonschema==4.1.0
MarkupSafe==2.0.1
numpy==1.21.4
pdf2image==1.16.0
Pillow==8.4.0
pymsteams==0.1.16
pyrsistent==0.18.0
pytesseract==0.3.8
pytz==2021.3
requests==2.26.0
six==1.16.0
urllib3==1.26.7
Werkzeug==2.0.2
